Create issue template for requesting access as a conference organizer
The readme says that we can grant access to this repo to conference organizer.

But it's not always clear whether someone is a conference organizer or not.

I created this issue template for requesting access to the repo. The idea is to allow a conference organizer to self identify of that they're an organizer and would like access to this project.