Port CSV linter to add new functionality
This change ports the linter from `csvclean` to `vladiate`. The primary
reason for this change is to add field-level validation to the existing
well-formatted / row length validation that `csvclean` provides. Three
types of field-level validation are included based on the documented
requirements in the readme:

1. Not-empty validation for required fields
2. Date format validation for isoformat dates
3. A set validator for ISO 3166-1 Alpha-3 country codes

A row length validator is also included to ensure that CSV files have
the correct number of fields and can be richly displayed on GitHub.